+++ packages/js/core/src/types/Invoke.ts
@@ -20,7 +20,7 @@ export interface InvokeApiOptions<
    * Input arguments for the method, structured as a map,
    * removing the chance of incorrectly ordering arguments.
    */
-  input: Record<string, unknown> | ArrayBuffer;
+  input?: Record<string, unknown> | ArrayBuffer;
 
   /**
    * Filters the [[InvokeApiResult]] data properties. The key
@@ -31,10 +31,10 @@ export interface InvokeApiOptions<
   resultFilter?: Record<string, unknown>;
 
   /**
-   * If set to true, the invoke function will decode all msgpack results
-   * into JavaScript objects.
+   * If set to true, the invoke function will not decode the msgpack results
+   * into JavaScript objects, and instead return the raw ArrayBuffer.
    */
-  decode?: boolean;
+  noDecode?: boolean;
 
   /**
    * Override the client's config for all invokes within this query.