+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Web3API 0.0.1-prealpha.58
+## Features
+* `@web3api/client-js`: Added `noDecode` invocation option.
+* `@web3api/client-js`: Added `noDefaults` constructor option.
+
+## Bugs
+* `@web3api/ethereum-plugin-js`: The `encodeParams` now properly parses arguments of type Array & Tuple.
+
 # Web3API 0.0.1-prealpha.57
 ## Features
 * `@web3api/cli`: CLI command middleware support has been added. The first use-cases implemented are to help ensure Docker is available to the CLI instance, and not in-use by another CLI instance.